GVT-3030: Fix V155 migration null cast for switch_in_layout_context

+++ infra/src/main/resources/db/migration/prod/V155__backfill_layout_switch_geometry_switch_id.sql
@@ -32,9 +32,9 @@ linked_switch_pairs as (
     select distinct le.geometry_switch_id, le.layout_switch_id
     from linked_edge le
         inner join layout.location_track_version_edge ltve on ltve.edge_id = le.edge_id
-        inner join layout.switch_in_layout_context('OFFICIAL'::layout.publication_state, null) ls
+        inner join layout.switch_in_layout_context('OFFICIAL'::layout.publication_state, null::int) ls
                    on ls.id = le.layout_switch_id
-        inner join layout.location_track_in_layout_context('OFFICIAL'::layout.publication_state, null) lt
+        inner join layout.location_track_in_layout_context('OFFICIAL'::layout.publication_state, null::int) lt
                    on ltve.location_track_id = lt.id
                      and ltve.location_track_layout_context_id = lt.layout_context_id
                      and ltve.location_track_version = lt.version
